Move types into categorized folders
Reorganized the `Domainic::Type` files into categorized subfolders to
improve clarity and scalability. The current categories are
`specification` and `core`, with the expectation that additional
categories will be introduced in the future.
