+++ domainic-type/lib/domainic/type/types/core/array_type.rb
