+++ domainic-type/lib/domainic/type/types/specification/anything_type.rb
