+++ domainic-type/lib/domainic/type/types/specification/duck_type.rb
