+++ domainic-type/lib/domainic/type/types/specification/union_type.rb
