+++ domainic-type/lib/domainic/type/types/specification/void_type.rb
