+++ domainic-type/sig/domainic/type/types/core/array_type.rbs
