+++ domainic-type/sig/domainic/type/types/specification/anything_type.rbs
