+++ domainic-type/sig/domainic/type/types/specification/duck_type.rbs
