+++ domainic-type/sig/domainic/type/types/specification/union_type.rbs
