+++ domainic-type/sig/domainic/type/types/specification/void_type.rbs
